Move SpringRepositoryFactoryBean from eventsourcing to modelling
Move SpringRepositoryFactoryBean from eventsourcing to modelling, since
repositories are not strictly event sourcing only. You can also have a
state-stored aggregate.

#2370
